pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pg-mirror"
version = "1.0.0"
description = "Ferramenta performática para espelhamento de bancos PostgreSQL com processamento paralelo"
readme = "README.md"
authors = [
    {name = "Seu Nome", email = "seu.email@exemplo.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: System :: Archiving :: Backup",
]
keywords = ["postgresql", "backup", "restore", "database", "migration", "mirror", "cli"]
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "pyjwt>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pylint>=2.15",
    "black>=22.0",
    "isort>=5.10",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/seu-usuario/pg-mirror"
Documentation = "https://github.com/seu-usuario/pg-mirror#readme"
Repository = "https://github.com/seu-usuario/pg-mirror.git"
Issues = "https://github.com/seu-usuario/pg-mirror/issues"

[project.scripts]
pg-mirror = "pg_mirror.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["pg_mirror*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pg_mirror --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "R0913",  # Too many arguments
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
